+++ DomainManagement/changelog.md
@@ -3,6 +3,7 @@
 ## ???
 
 - Upd: Component Group Membership - now can define Group Processing Mode, introducing Constrained and Additive Modes to a group's memberships.
+- Upd: Component Group Membership - configured names may now include SIDs, such as the SIDs of built-in accounts or groups.
 - Fix: Register-DMGroupMembership - explicitly registering empty as $false no longer clears configured settings.
 
 ## 1.3.70 (2020-06-12)