+++ DomainManagement/changelog.md
@@ -1,6 +1,6 @@
 ﻿# Changelog
 
-## ???
+## 1.3.75 (2020-07-31)
 
 - Upd: Component Group Membership - now can define Group Processing Mode, introducing Constrained and Additive Modes to a group's memberships.
 - Upd: Component Group Membership - configured names may now include SIDs, such as the SIDs of built-in accounts or groups.